refactor: modularize formatting functions
