+++ js/index.js
@@ -26,7 +26,14 @@ const summarizePlayerData = require('./js/database/summarize-player-data');
 const summarizeTalentData = require('./js/database/summarize-talent-data');
 const summarizeTeamData = require('./js/database/summarize-team-data');
 
-const migrateDatabase = require('./js/database/migrate.js');
+const migrateDatabase = require('./js/database/migrate');
+
+const {
+  formatSeconds,
+  formatStat,
+  capitalize
+} = require('./js/util/formatters');
+
 
 const RegionString = {
   1: 'NA',
@@ -321,38 +328,6 @@ function setMenuTitle(title, showBackButton) {
   }
 }
 
-// formats to mm:ss
-function formatSeconds(val) {
-  let invert = false;
-  if (val < 0)
-    invert = true;
-  let fval = Math.abs(val);
-
-  let duration = new Date(fval * 1000);
-  let seconds = duration.getUTCSeconds();
-  let minutes = duration.getUTCMinutes();
-
-  return (invert ? '-' : '') + (minutes < 1 ? "0" : "") + minutes + ":" + (seconds < 10 ? "0" : "") + seconds;
-}
-
-function formatStat(field, val, allFixed = false) {
-  if (val === undefined)
-    return 0;
-
-  if (field === 'KillParticipation' || field === 'timeDeadPct' || field === 'mercUptimePercent' || field === 'pct')
-    return (val * 100).toLocaleString(undefined, { maximumFractionDigits: 1}) + '%';
-  else if (field === 'KDA')
-    return val.toLocaleString(undefined, { maximumFractionDigits: 1});
-  else if (field.startsWith('Time') || field === 'OnFireTimeOnFire' || field === 'timeTo10' ||
-    field === 'timeTo20' || field === 'mercUptime' || field === 'avgTimeSpentDead')
-    return formatSeconds(val);
-
-  if (allFixed) {
-    return val.toLocaleString(undefined, { maximumFractionDigits: 1});
-  }
-
-  return val.toLocaleString();
-}
 
 // updates certain elements based on a new replay inserted into the database
 function globalDBUpdate() {
@@ -428,10 +403,6 @@ function addMapMenuOptions(menu) {
     let map = ReplayTypes.MapType[m];
     menu.find('.menu').append('<div class="item" data-value="' + map + '">' + map + '</div>');
   }
-}
-
-function capitalize(string) {
-  return string.charAt(0).toUpperCase() + string.slice(1);
 }
 
 function addPatchMenuOptions(elem, callback) {
